requirements.txt
# ── WellBeing Agent Dependencies ──────────────────────────────────────

# Web framework
fastapi
uvicorn[standard]
python-multipart
pydantic

# HTTP client
httpx

# Environment / config
python-dotenv

# Language detection
langdetect

# RAG pipeline (LlamaIndex)
llama-index-core
llama-index-embeddings-huggingface
sentence-transformers

# Hugging Face
huggingface_hub

# Whisper Large v3 (speech-to-text)
transformers>=4.36.0
torch
accelerate

# Numerical
numpy